Fix tests for ActivationResult + should_fake_success (#81)
* fix should_fake_success and add tests for it.

Previously, should_fake_success was always returning True if
WAKEPY_FAKE_SUCCESS was set to any value as for example bool('0') is
True.

* fix tox tests (missing jeepney)

The tox tests were missing a requirement (jeepney)

* fix tests for activationresult

The test logic was broken in one case, and WAKEPY_FAKE_SUCCESS was not
set correctly for the test. No functional changes.