Pin dev dependencies (reproducible tests & doc builds) (#220)
Pin all development dependencies. Well, except for IPython which really
is just there as convenience tool and not used within the repo.

Tests and building docs should be reproducible.

Python 3.7 gets special treatment as the support for it is dropped from
many packages. I don't want to use older version of tools day-to-day and
with every version of python just because of python 3.7. Hence, it has
different versions pinned. If that becomes a problem, have to use the
same (older) version also when testing with other python versions.