Run mypy for full repo & fix all mypy issues (#227)
mypy
----

Previously, mypy was only ran for ./wakepy folder. But there seems some
added value in running static type checking also for tests, since it
also shows how different parts of wakepy look like when running mypy on
a project that uses wakepy. I found at least places which were not so
user friendly for people using wakepy, and fixed them. I also fixed
80+ mypy errors.

StrEnums and their values
-------------------------

Each subclass of StrEnum has not corresponding Values class. For example
PlatformName has PlatformNameValues which is a typing.Literal. This is
to allow giving the literal values (strings) as function arguments, when
a function expects value listed in StrEnum. The StrEnum and the
corresponding Values are enforced to be in sync with a unit test.

StrEnum class is now simplified a bit.


do_assert
---------
added a do_assert helper for tests which can be used in place of assert
statement if you're getting [unreachable] errors with mypy.

Other
----
Mode.activation_result is now ActivationResult (was: ActivationResult |
None). This makes it easier to work with the object when using wakepy and
  mypy in a project.

Method.name is now always a string. Unnamed methods have special name
"__unnamed__". This makes working with wakepy+mypy easier. There is also
now Method._is_unnamed() which can be used to check if the Method has a
name.