Add publishing workflow (GitHub actions) (#238)
Add a Github workflow file which enables automatic signing and
publishing to PyPI and to GitHub Releases when a tag is added to a
commit on latest-release branch. This is mostly tested but might require
some small adjustments when doing the first release (0.8.0) using this.

The publishing workfow uses the build and test workflow.