mypy: add warn_unused_ignores=true
Also removed the skip_intall=True in tox 'check' environment. (this made
things easier as wakepy.__init__ requires wakepy._version to be present
which is created in a build step, which is only done if skip_install is
not True, and if the wakepy._version is not available, mypy would complain
about it. And if we ignore the import line, mypy will next complain about
unused ignores (when the wakepy._version file is present).