Remove MethodError (#269)
The usage of MethodError does not seem systematic, and it is not clear
when to use it. Replaced MethodError in the few places it was used with
RuntimeError (MethodError's superclass).

Could add a MethodError or similar back if there is clear need for it.
But for now, keeping it simple.