Make on_fail="warn" default (was: "error") (#388)
Closes: #376

This change can also be reverted if this seems to be troublesome. But
with default on_fail of "warn", most people will be probably happier and
do not think that wakepy is "flaky" e.g. on linux, where it's not always
possible (at least yet) to prevent automatic idle action / suspend.