Fix prioritized order of Methods (#429)
Fixes: #428 

Description of the fixed bug: Sometimes wakepy tried different Methods
in the wrong order. For example, it could try the "caffeinate" method
(works only on MacOS) on Linux before any Linux-specific method. This
problem was introduced in https://github.com/fohrloop/wakepy/pull/379/
and was present in wakepy versions 0.10.0 to 0.10.2.