Fix flaky tests on pypy+windows (#447)
Fixes: #443

Background: Sometimes tests failed in the pipelines with PyPy+Windows combination.

The explanation for this problem was pretty simple:
- When using wakepy normally, it waits undefinitely in the worker thread
for uninhibit command (release event)
- During tests, a 1ms timeout value was used for the release event. This
was enough in every other case, but not with PyPy+Windows combination.
For some reason, code runs slower with PyPy+Windows and the timeout
needed to be increased.

As a fix: Changed the timeout from 1ms to 500ms. Also added a helpful(?)
warning which will be issued if the queue gets two items instead of one
(which happens if the timeout is too small).