Make Run Tox fail on Windows pipelines if pytest fails (#442)
Fixes: #441 

The pipeline runFixes: #441 

The pipeline run 16274295300/job/45949378559 demonstrates a successful
failure.

Without the `set -e` and the `shell: bash`, on windows the last command,
which is the `tox -e mypy` would determine the exit code of the step.
(the default behavior of multi-line run commands is different on linux
and windows)
[16274295300/job/45949378559](https://github.com/fohrloop/wakepy/actions/runs/16274295300/job/45949378559?pr=442)
demonstrates a successful failure.

The problem was in the workflow .yml file:

```
      - name: Run tox
        shell: bash
        run: |
          set -e
          python -m pip install ${{ env.TOX_REQUIREMENT }}
          tox -e ${{ matrix.python-version }} --skip-build
          tox -e mypy --skip-build
```

Without the `set -e` and the `shell: bash`, on windows the last command,
which is the `tox -e mypy` would determine the exit code of the step.
(the default behavior of multi-line run commands is different on linux
and windows)